add post samples to HomePage + Comment Comp
