+++ frontend/src/component/comment_input/Comment_Input.css
@@ -1,0 +1,12 @@
+.comment_input input{
+    width: 80%;
+    padding: 4px 4px;
+    margin: 8px 0;
+    box-sizing: border-box;
+    border: 3px solid #000000;
+    border-radius: 5px;
+}
+
+.comment_input input:focus{
+    border: 3px solid #2ecc71;
+}
