+++ frontend/src/pages/HomePage.js
@@ -1,5 +1,9 @@
 import style from "./HomePage.module.css";
 
+// just for showing posts
+import { useFetch } from '../hooks/useFetch'
+import Postlist from '../component/postlist/Postlist'
+
 
 // import { useAuthContext } from "../hooks/useAuthContext";
 import { useEffect, useState } from "react";
@@ -20,6 +24,10 @@ export default function HomePage() {
   window.onbeforeunload = function (event) {
     document.body.classList.remove(style.bodyclass);
   };
+
+
+  // just for posts
+  const { data , loading , error } = useFetch('http://localhost:3000/posts')
   
   const navigator = useNavigate();
 
@@ -48,6 +56,13 @@ export default function HomePage() {
       </div>
       <ShowPlaceHomePage />
       <AuctionsHomePage />
+ 
+       {/* posts */}
+      <div className='home'>
+        {error && <p className='error'>{error}</p>}
+        {loading && <p className='loading'>Loading...</p>}
+        {data && <Postlist posts={data} />}
+    </div>
     </div>
   );
 }
