+++ frontend/src/pages/PS_Artist.js
@@ -11,6 +11,11 @@ import HeaderProfile from "../sections/HeaderProfile";
 import addp1 from "../img/addpost1.png";
 import addp2 from "../img/addpost2.png";
 import { useEffect , useState} from "react";
+
+// just for showing posts
+import { useFetch } from '../hooks/useFetch'
+import Postlist from '../component/postlist/Postlist'
+
 export default function PS_Artist() {
   
   document.body.className = '';
@@ -19,6 +24,9 @@ export default function PS_Artist() {
   window.onbeforeunload = () => {
     document.body.classList.remove(style.bodyclass);
   };
+
+  // just for posts
+  const { data , loading , error } = useFetch('http://localhost:3000/posts')
 
   const navigator = useNavigate();
 
@@ -51,6 +59,14 @@ export default function PS_Artist() {
         </div>)}
       <ShowPlaceProfile artistId={artistId} />
       <PostProfile artistId={artistId} />
+
+       {/* posts */}
+       <div className='home'>
+          {error && <p className='error'>{error}</p>}
+          {loading && <p className='loading'>Loading...</p>}
+          {data && <Postlist posts={data} />}
+        </div>
+    
     </div>
   );
 }
