+++ frontend/src/sections/AuctionsHomePage.js
@@ -114,7 +114,7 @@ export default function AuctionsHomePage() {
       <div id={style.asara}>
         <button className={style.blue}>آثار هنری</button>
       </div>
-      <img src={posts} alt="posts" id={style.bpost} />
+      
     </div>
   );
 }
